backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.0
pydantic==2.8.2
python-dotenv==1.0.1
SQLAlchemy==2.0.32
sqlmodel==0.0.21
apscheduler==3.10.4
firebase-admin==6.5.0
requests==2.32.4
aiohttp==3.9.5
pandas==2.3.2
numpy==2.3.2
scikit-learn==1.7.1
joblib==1.5.1
python-multipart==0.0.9
PyPDF2==3.0.1
python-docx==1.1.2
Pillow==11.3.0
beautifulsoup4==4.13.4
google-generativeai
langchain-google-genai
langchain
huggingface_hub
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
flask==3.0.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
outcome==1.3.0.post0
prettytable==3.9.0
pycparser==2.21
PySocks==1.7.1
PyYAML==6.0.1
selenium==4.15.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.0.7
wcwidth==0.2.9
werkzeug==3.0.1
wsproto==1.2.0
zipp==3.17.0
gunicorn==20.1.0
